アーキテクチャの変更(Domain層を導入)
