+++ jobHunt/Domain/Model/Event.swift
