+++ jobHunt/Domain/Model/InterView.swift
