+++ jobHunt/Domain/Model/Internship.swift
