+++ jobHunt/Domain/Model/Session.swift
