+++ jobHunt/Domain/Repository/FormatRepository.swift
@@ -1,5 +1,5 @@
 //
-//  FormatModel.swift
+//  FormatRepository.swift
 //  jobHunt
 //
 //  Created by 出口楓真 on 2022/10/27.
@@ -7,10 +7,10 @@
 
 import Foundation
 
-class FormatModel {
-    let dateFormatter = DateFormatter()
+class FormatRepository {
 
     func format(date: Date) -> String {
+        let dateFormatter = DateFormatter()
         dateFormatter.dateFormat = "YYYY年MM月dd日"
         return dateFormatter.string(from:date)
     }
