+++ jobHunt/Feature/Detail/DetailView 2.swift
