+++ jobHunt/Feature/Detail/DetailView.swift
