+++ jobHunt/Feature/Detail/ESDetailView 2.swift
