+++ jobHunt/Feature/Detail/ESDetailView.swift
