+++ jobHunt/Feature/Detail/InternshipDetailView.swift
