+++ jobHunt/Feature/Home/HomeModel.swift
@@ -1,0 +1,19 @@
+//
+//  HomeModel.swift
+//  jobHunt
+//
+//  Created by 出口楓真 on 2022/10/14.
+//
+
+import Foundation
+
+class HomeModel {
+
+    var esArray: [ES] = []
+    var interviewArray: [Interview] = []
+    var sessionArray: [Session] = []
+    var internshipArray: [Internship] = []
+    var events: [any Entry] = []
+
+}
+
